Fixing errors found by `cargo clippy`
