+++ src/lib.rs
@@ -1,3 +1,5 @@
+#![allow(non_snake_case)]
+
 // Declare the modules that form the library's public API (or internal structure)
 // Using `pub mod` makes them accessible from the binaries using `use rust_data::module_name;`
 pub mod config;
